Cleanup dafavorites package, isolate marshaling types.
Dedicate dafavorites package to the CLI and nothing else. That was
probably the original intent because in general it's a good idea to
isolate CLI and then the rest of the code should be oblivious about it.
Easier to replace the CLI with a GUI or something else.

Isolate all types involved in marshaling or unmarshaling. Otherwise you
have to keep guessing whether whether the types and their properties
have to be exported (i.e. capitalized).

Finally, unexport those functions and types that don't anymore have to
be.